src/requirements.txt
Django>=6.0.1,<7.0
djangorestframework>=3.16.0,<4.0
pytest>=8.0
pytest-django>=4.8

pytest==8.*
pytest-django==4.*
factory-boy==3.*
